pyproject.toml
[tool.poetry]
name = "my-fastapi-template"
version = "0.1.0"
description = ""
authors = ["artyomklg <artem0801200269@gmail.com>"]
readme = "README.md"
packages = [{include = "my_fastapi_template"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.100.0"}
sqlalchemy = "^2.0.18"
asyncpg = "^0.27.0"
alembic = "^1.11.1"
gunicorn = "^20.1.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
flake8 = "^6.0.0"
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
